pyproject.toml
[project]
name = "cogesak-lakehouse-mapper"
version = "0.1.0"
description = "A simple 3-layer architecture project"
authors = [
    { name = "Esakki" }
]
requires-python = ">=3.12"


dependencies = [
    "delta-spark==3.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",

]

[tools.ruff]
exclude = ["tests"]
line-length = 122
target-version="py312"



[tool.ruff.lint]
select = [
#    "F",
    "I",
    "W291",
    "F401"
]
